+++ modules/excel/src/workers/excel-loader.worker.js
@@ -1,0 +1,5 @@
+import {createWorker} from '@loaders.gl/loader-utils';
+
+import {ExcelLoader} from '../excel-loader';
+
+createWorker(ExcelLoader);